Strip modern toolchain features
The xv6 ELF handling is horribly underpowered, it breaks on all sorts of
modern toolchain features.  A prime example of this is the various
GNU-specific note sections that GCC emits.  We can strip those out.

Similarly, modern GCC heavily uses CPU extensions not enabled in the
context of xv6.  To suppress this effect, we can use
-mgeneral-regs-only.

GNU ld also seems to adjust PAs to be equal to VAs if we manually move
'.', so, instead, we can tell .data to be aligned to a page, which fixes
the kernel failing to map itself.